Combine the ring functionality directly into the CacheRing since it's so tightly coupled. Remove the vestigial DequeRing.
